split range input stories
